refactor(api): extract summary recalculation into a shared service
