+++ apps/api/src/tasks/summary-recalculation.service.spec.ts
@@ -1,0 +1,8 @@
+import { SummaryRecalculationService } from './summary-recalculation.service';
+
+describe('SummaryRecalculationService', () => {
+  it('exposes a recalculate method', () => {
+    const svc = new SummaryRecalculationService({} as any);
+    expect(typeof svc.recalculate).toBe('function');
+  });
+});
