+++ apps/api/src/tasks/tasks.service.ts
@@ -27,8 +27,6 @@ import {
   UpdateTaskPositionDto,
   UpdateProgressDto,
   UpdateTaskDto,
-  computeSummaryBounds,
-  SummaryBoundsNode,
 } from '@project-workgroup/shared';
 import { AuthUser } from '../auth/auth.guard';
 import {
@@ -46,9 +44,7 @@ import {
   SchedulingService,
   ScheduledSlot,
 } from '../calendar/scheduling.service';
-
-type TxClient = Prisma.TransactionClient;
-type DbClient = PrismaService | TxClient;
+import { SummaryRecalculationService } from './summary-recalculation.service';
 
 type TaskRow = {
   id: bigint;
@@ -75,25 +71,6 @@ type TaskRow = {
   updatedAt: Date;
 };
 
-type SummaryCalcRow = Pick<
-  TaskRow,
-  | 'id'
-  | 'parentId'
-  | 'startDate'
-  | 'endDate'
-  | 'duration'
-  | 'progress'
-  | 'type'
-  | 'estimatedHours'
->;
-type SummaryStats = {
-  weightedProgress: number;
-  progressWeight: number;
-  fallbackProgress: number;
-  fallbackCount: number;
-  estimatedHours: number;
-};
-
 type TaskUpdatePlan = {
   scheduleTouched: boolean;
   workloadTouched: boolean;
@@ -107,6 +84,7 @@ export class TasksService {
     private readonly prisma: PrismaService,
     private readonly calendarResolver: CalendarResolverService,
     private readonly scheduling: SchedulingService,
+    private readonly summaries: SummaryRecalculationService,
     @Optional()
     @InjectPinoLogger(TasksService.name)
     private readonly logger?: PinoLogger,
@@ -525,185 +503,6 @@ export class TasksService {
     }
   }
 
-  private async recalculateProjectSummaries(
-    projectId: bigint,
-    client: DbClient = this.prisma,
-  ): Promise<SummaryPatch[]> {
-    const tasks = await client.task.findMany({
-      where: { projectId },
-      select: {
-        id: true,
-        parentId: true,
-        startDate: true,
-        endDate: true,
-        duration: true,
-        progress: true,
-        type: true,
-        estimatedHours: true,
-      },
-    });
-
-    const childrenByParent = new Map<string, SummaryCalcRow[]>();
-    for (const task of tasks) {
-      if (task.parentId === null) continue;
-      const key = task.parentId.toString();
-      const list = childrenByParent.get(key);
-      if (list) list.push(task);
-      else childrenByParent.set(key, [task]);
-    }
-
-    // Fechas (start/end) de cada summary: regla compartida con el Gantt web (packages/shared).
-    const boundsMap = computeSummaryBounds(
-      tasks.map(
-        (t): SummaryBoundsNode => ({
-          id: t.id.toString(),
-          parentId: t.parentId?.toString() ?? null,
-          type: t.type,
-          start: t.startDate?.getTime() ?? null,
-          end: t.endDate?.getTime() ?? null,
-        }),
-      ),
-    );
-
-    const memo = new Map<string, SummaryStats>();
-
-    // Progreso y horas agregados del subárbol (las fechas vienen de boundsMap).
-    const collect = (task: SummaryCalcRow): SummaryStats => {
-      const key = task.id.toString();
-      const cached = memo.get(key);
-      if (cached) return cached;
-
-      const children = childrenByParent.get(key) ?? [];
-      let weightedProgress = 0;
-      let progressWeight = 0;
-      let fallbackProgress = 0;
-      let fallbackCount = 0;
-      let estimatedHours = 0;
-
-      if (children.length === 0) {
-        if (task.type !== 'summary') {
-          const weight = Number(task.duration.toString());
-          if (weight > 0) {
-            weightedProgress += weight * task.progress;
-            progressWeight += weight;
-          } else {
-            fallbackProgress += task.progress;
-            fallbackCount++;
-          }
-          const ownHours = Number(task.estimatedHours.toString());
-          if (Number.isFinite(ownHours) && ownHours > 0)
-            estimatedHours += ownHours;
-        }
-      } else {
-        for (const child of children) {
-          const childStats = collect(child);
-          weightedProgress += childStats.weightedProgress;
-          progressWeight += childStats.progressWeight;
-          fallbackProgress += childStats.fallbackProgress;
-          fallbackCount += childStats.fallbackCount;
-          estimatedHours += childStats.estimatedHours;
-        }
-      }
-
-      const result = {
-        weightedProgress,
-        progressWeight,
-        fallbackProgress,
-        fallbackCount,
-        estimatedHours,
-      };
-      memo.set(key, result);
-      return result;
-    };
-
-    const patches: SummaryPatch[] = [];
-    const previousById = new Map<
-      string,
-      {
-        startDate: Date;
-        endDate: Date;
-        duration: string;
-        progress: number;
-        estimatedHours: string;
-      }
-    >();
-    for (const task of tasks) {
-      if (task.type === 'summary') {
-        previousById.set(task.id.toString(), {
-          startDate: task.startDate,
-          endDate: task.endDate,
-          duration: task.duration.toString(),
-          progress: task.progress,
-          estimatedHours: task.estimatedHours.toString(),
-        });
-      }
-    }
-
-    for (const summary of tasks.filter((task) => task.type === 'summary')) {
-      const stats = collect(summary);
-      const bounds = boundsMap.get(summary.id.toString());
-      if (!bounds) continue;
-      const startDate = new Date(bounds.start);
-      const endDate = new Date(bounds.end);
-
-      const progress =
-        stats.progressWeight > 0
-          ? Math.round(stats.weightedProgress / stats.progressWeight)
-          : stats.fallbackCount > 0
-            ? Math.round(stats.fallbackProgress / stats.fallbackCount)
-            : 0;
-      const clampedProgress = Math.max(0, Math.min(100, progress));
-      const duration = Math.max(
-        0,
-        Math.ceil(
-          (endDate.getTime() - startDate.getTime()) / (1000 * 60 * 60 * 24),
-        ),
-      ).toString();
-
-      const aggregatedHours = stats.estimatedHours.toFixed(2);
-      const prev = previousById.get(summary.id.toString());
-      const unchanged =
-        prev &&
-        prev.startDate.getTime() === startDate.getTime() &&
-        prev.endDate.getTime() === endDate.getTime() &&
-        prev.duration === duration &&
-        prev.progress === clampedProgress &&
-        Number(prev.estimatedHours) === Number(aggregatedHours);
-      if (unchanged) continue;
-
-      const updated = await client.task.update({
-        where: { id: summary.id },
-        data: {
-          startDate,
-          endDate,
-          duration,
-          progress: clampedProgress,
-          estimatedHours: new Prisma.Decimal(aggregatedHours),
-          version: { increment: 1 },
-        },
-        select: {
-          id: true,
-          startDate: true,
-          endDate: true,
-          duration: true,
-          progress: true,
-          estimatedHours: true,
-          version: true,
-        },
-      });
-      patches.push({
-        id: updated.id.toString(),
-        startDate: updated.startDate.toISOString(),
-        endDate: updated.endDate.toISOString(),
-        duration: updated.duration.toString(),
-        progress: updated.progress,
-        estimatedHours: updated.estimatedHours.toString(),
-        version: updated.version,
-      });
-    }
-    return patches;
-  }
-
   async list(projectId: bigint): Promise<TaskResponse[]> {
     const [tasks, calendar] = await Promise.all([
       this.prisma.task.findMany({
@@ -759,7 +558,7 @@ export class TasksService {
             schedule.workload,
           );
         }
-        await this.recalculateProjectSummaries(projectId, tx);
+        await this.summaries.recalculate(projectId, tx);
         return tx.task.findUnique({ where: { id: created.id } }) ?? created;
       },
       {
@@ -877,7 +676,7 @@ export class TasksService {
             );
           }
           const patched = plan.summariesTouched
-            ? await this.recalculateProjectSummaries(existing.projectId, tx)
+            ? await this.summaries.recalculate(existing.projectId, tx)
             : [];
           const fresh =
             (await tx.task.findUnique({ where: { id } })) ?? updated;
@@ -935,7 +734,7 @@ export class TasksService {
             where,
             data: { progress: dto.progress, version: { increment: 1 } },
           });
-          const patched = await this.recalculateProjectSummaries(
+          const patched = await this.summaries.recalculate(
             existing.projectId,
             tx,
           );
@@ -1071,7 +870,7 @@ export class TasksService {
             },
           });
           const patched = parentChanged
-            ? await this.recalculateProjectSummaries(existing.projectId, tx)
+            ? await this.summaries.recalculate(existing.projectId, tx)
             : [];
           return { task, patched };
         },
@@ -1194,7 +993,7 @@ export class TasksService {
     await this.prisma.$transaction(
       async (tx) => {
         await tx.task.delete({ where: { id } });
-        await this.recalculateProjectSummaries(existing.projectId, tx);
+        await this.summaries.recalculate(existing.projectId, tx);
       },
       {
         isolationLevel: Prisma.TransactionIsolationLevel.ReadCommitted,
@@ -1480,7 +1279,7 @@ export class TasksService {
         }
 
         const summariesPatched = summariesDirty
-          ? await this.recalculateProjectSummaries(projectId, tx)
+          ? await this.summaries.recalculate(projectId, tx)
           : [];
         return { tasks: updatedTasks, summariesPatched };
       },
